feat!: clean up deprecated Databases and DataTypes, remove migrations from old versions (+1yr old). Direct data migrations will now only be supported in a 12 month window.
