--- meteor/server/migration/0_18_0.ts
@@ -1,4 +1,0 @@
-import { addMigrationSteps } from './databaseMigration'
-
-// 0.18.0: Release 4
-export const addSteps = addMigrationSteps('0.18.0', [])
