--- meteor/server/migration/0_21_0.ts
@@ -1,4 +1,0 @@
-import { addMigrationSteps } from './databaseMigration'
-
-// 0.21.0 ( Release 6 )
-export const addSteps = addMigrationSteps('0.21.0', [])
