--- meteor/server/migration/0_22_0.ts
@@ -1,4 +1,0 @@
-import { addMigrationSteps } from './databaseMigration'
-
-// 0.22.0 (Release 7)
-export const addSteps = addMigrationSteps('0.22.0', [])
