--- meteor/server/migration/1_10_0.ts
@@ -1,4 +1,0 @@
-import { addMigrationSteps } from './databaseMigration'
-
-// Release 22
-export const addSteps = addMigrationSteps('1.10.0', [])
