--- meteor/server/migration/1_16_0.ts
@@ -1,27 +1,0 @@
-import { addMigrationSteps } from './databaseMigration'
-
-/*
- * **************************************************************************************
- *
- *  These migrations are destined for the next release
- *
- * (This file is to be renamed to the correct version number when doing the release)
- *
- * **************************************************************************************
- */
-// Release 28
-export const addSteps = addMigrationSteps('1.16.0', [
-	//                     ^--- To be set to an absolute version number when doing the release
-	// add steps here:
-	// {
-	// 	id: 'my fancy step',
-	// 	canBeRunAutomatically: true,
-	// 	validate: () => {
-	// 		return false
-	// 	},
-	// 	migrate: () => {
-	// 		//
-	// 	}
-	// },
-	//
-])
