+++ meteor/server/migration/1_19_0.ts
@@ -2,14 +2,7 @@ import { CollectionName } from '@sofie-automation/corelib/dist/dataModel/Collect
 import { addMigrationSteps } from './databaseMigration'
 import { ensureCollectionProperty } from './lib'
 
-/*
- * **************************************************************************************
- *
- *  These migrations are destined for the next release
- *
- * **************************************************************************************
- */
-// Release 31
+// Release 31 (2021-05-05)
 export const addSteps = addMigrationSteps('1.19.0', [
 	ensureCollectionProperty(CollectionName.CoreSystem, {}, 'cron.casparCGRestart.enabled', true),
 ])
