+++ meteor/server/migration/1_33_0.ts
@@ -1,4 +1,4 @@
 import { addMigrationSteps } from './databaseMigration'
 
-// Release 33
+// Release 33 (2021-06-16)
 export const addSteps = addMigrationSteps('1.33.0', [])
