+++ meteor/server/migration/1_34_0.ts
@@ -1,4 +1,4 @@
 import { addMigrationSteps } from './databaseMigration'
 
-// Release 34
+// Release 34 (2021-06-30)
 export const addSteps = addMigrationSteps('1.34.0', [])
