+++ meteor/server/migration/1_38_0.ts
@@ -2,16 +2,7 @@ import { addMigrationSteps } from './databaseMigration'
 import { TriggeredActions } from '../../lib/collections/TriggeredActions'
 import { getHash, protectString, unprotectString } from '../../lib/lib'
 
-/*
- * **************************************************************************************
- *
- *  These migrations are destined for the next release
- *
- * (This file is to be renamed to the correct version number when doing the release)
- *
- * **************************************************************************************
- */
-// Release 38
+// Release 38 (2022-01-27)
 export const addSteps = addMigrationSteps('1.38.0', [
 	// Add some migrations!
 	{
