--- meteor/server/migration/1_3_0.ts
@@ -1,4 +1,0 @@
-import { addMigrationSteps } from './databaseMigration'
-
-// 1.3.0 (Release 15)
-export const addSteps = addMigrationSteps('1.3.0', [])
