--- meteor/server/migration/1_4_0.ts
@@ -1,4 +1,0 @@
-import { addMigrationSteps } from './databaseMigration'
-
-// 1.4.0 (Release 16)
-export const addSteps = addMigrationSteps('1.4.0', [])
