--- meteor/server/migration/1_6_0.ts
@@ -1,4 +1,0 @@
-import { addMigrationSteps } from './databaseMigration'
-
-// 1.6.0 (Release 18)
-export const addSteps = addMigrationSteps('1.6.0', [])
