--- meteor/server/migration/1_8_0.ts
@@ -1,4 +1,0 @@
-import { addMigrationSteps } from './databaseMigration'
-
-// 1.8.0 (Release 20)
-export const addSteps = addMigrationSteps('1.8.0', [])
